+++ pkg/sopsenv/error.go
