+++ pkg/vaultclient/error.go
