+++ pkg/vaultclient/vaultclient.go
